creates rest-jira-exporter.py
very incomplete, but a good start. Able to get all of the issues into a dictionary. Still need to format that into a csv for re-import so this could be used to "migrate" a project from one jira to another given only access to the 1 project and the REST API of whatever environment being migrated away from.